+++ src/app/search/search.controller.js
@@ -65,6 +65,9 @@
 
         vm.defaultRefineModel = {
             acb: {
+                'CCHIT': false,
+                'SLI Global': false,
+                'Surescripts LLC': false,
                 'Drummond Group': true,
                 'ICSA Labs': true,
                 'InfoGard': true
@@ -84,6 +87,9 @@
                 'Suspended by ONC': true,
                 'Terminated by ONC': false
             }
+        };
+        vm.retired = {
+            acb: {'CCHIT': true, 'SLI Global': true, 'Surescripts LLC': true}
         };
 
         function browseAll () {
@@ -372,7 +378,7 @@
 
         function populateSearchOptions () {
             vm.lookaheadSource = {all: [], developers: [], products: []};
-            commonService.getSearchOptions()
+            commonService.getSearchOptions(true)
                 .then(function (options) {
                     if (vm.isPreLoading) {
                         cfpLoadingBar.start();
@@ -383,6 +389,12 @@
                     options.practiceTypes = [];
                     for (i = 0; i < options.practiceTypeNames.length; i++) {
                         options.practiceTypes.push(options.practiceTypeNames[i].name);
+                    }
+                    for (i = 0; i < options.certBodyNames.length; i++) {
+                        if (options.certBodyNames[i].name === 'Pending') {
+                            options.certBodyNames.splice(i,1);
+                            break;
+                        }
                     }
                     for (i = 0; i < options.certificationStatuses.length; i++) {
                         if (options.certificationStatuses[i].name === 'Pending') {
@@ -429,7 +441,7 @@
         }
 
         function setFilterInfo (refineModel) {
-            var i;
+            var i, obj;
             vm.refineModel = angular.copy(refineModel);
             vm.filterItems = {
                 pageSize: '50',
@@ -441,7 +453,14 @@
             };
             vm.searchOptions.certBodyNames = $filter('orderBy')(vm.searchOptions.certBodyNames, 'name');
             for (i = 0; i < vm.searchOptions.certBodyNames.length; i++) {
-                vm.filterItems.acbItems.push({value: vm.searchOptions.certBodyNames[i].name, selected: vm.defaultRefineModel.acb[vm.searchOptions.certBodyNames[i].name]});
+                obj = {
+                    value: vm.searchOptions.certBodyNames[i].name,
+                    selected: vm.defaultRefineModel.acb[vm.searchOptions.certBodyNames[i].name]
+                };
+                if (vm.retired.acb[vm.searchOptions.certBodyNames[i].name]) {
+                    obj.display = obj.value + ' (Retired)';
+                }
+                vm.filterItems.acbItems.push(obj);
             }
             vm.searchOptions.editions = $filter('orderBy')(vm.searchOptions.editions, 'name');
             for (i = 0; i < vm.searchOptions.editions.length; i++) {